+++ staging/src/k8s.io/component-helpers/BUILD
@@ -9,9 +9,9 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
+        "//staging/src/k8s.io/component-helpers/apimachinery/lease:all-srcs",
         "//staging/src/k8s.io/component-helpers/auth/rbac/reconciliation:all-srcs",
         "//staging/src/k8s.io/component-helpers/auth/rbac/validation:all-srcs",
-        "//staging/src/k8s.io/component-helpers/lease:all-srcs",
         "//staging/src/k8s.io/component-helpers/scheduling/corev1:all-srcs",
     ],
     tags = ["automanaged"],
