+++ staging/src/k8s.io/component-helpers/apimachinery/OWNERS
@@ -1,0 +1,8 @@
+# See the OWNERS docs at https://go.k8s.io/owners
+
+approvers:
+- sig-api-machinery-api-approvers
+reviewers:
+- sig-api-machinery-api-reviewers
+labels:
+- sig/api-machinery
