+++ staging/src/k8s.io/component-helpers/apimachinery/lease/BUILD
@@ -3,8 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["controller.go"],
-    importmap = "k8s.io/kubernetes/vendor/k8s.io/component-helpers/lease",
-    importpath = "k8s.io/component-helpers/lease",
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/component-helpers/apimachinery/lease",
+    importpath = "k8s.io/component-helpers/apimachinery/lease",
     visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/api/coordination/v1:go_default_library",
